Added Debug modes
added debug distance mode which displays the velocity, position, and distance to the closest pickup. It also draw a line between the player and the closest pickup and turns it blue.
Also added debus vision mode, which shows the velocity vector visually, colours the pickup that the player is heading towards green, and make the pickup face towards the player.